more code improvements and some locale

+++ OOps/str_ops.c
@@ -166,10 +166,10 @@ int strcpy_opcode_S(CSOUND *csound, STRCPY_OP *p)
         mfree(csound, p->r->data);
         p->r->data = cs_strdup(csound, newVal);
         p->r->size = strlen(newVal) + 1;
-   
+
     }
     else strcpy((char*) p->r->data, newVal);
-        
+
     return OK;
 }
 
@@ -242,8 +242,8 @@ int strcat_opcode(CSOUND *csound, STRCAT_OP *p)
 {
 ;
   int size;
-  
-  size = strlen(p->str1->data) + 
+
+  size = strlen(p->str1->data) +
          strlen(p->str2->data);
 
     if(p->r->data == NULL) {
@@ -349,16 +349,16 @@ sprintf_opcode_(CSOUND *csound,
         case 'u':
         case 'c':
 #ifdef HAVE_SNPRINTF
-        if(strlen(strseg) + 24 > maxChars) {
-           int offs = outstring - str->data;
+          if ((int)strlen(strseg) + 24 > (int)maxChars) {
+            int offs = outstring - str->data;
             str->data = mrealloc(csound, str->data,
                                  str->size  + 13);
             str->size += 24;
             maxChars += 24;
             outstring = str->data + offs;
             //printf("maxchars = %d  %s\n", maxChars, strseg);
-	  }
-          n = snprintf(outstring, maxChars, strseg, (int) MYFLT2LRND(*parm));    
+          }
+          n = snprintf(outstring, maxChars, strseg, (int) MYFLT2LRND(*parm));
 #else
           n = sprintf(outstring, strseg, (int) MYFLT2LRND(*parm));
 #endif
@@ -370,7 +370,7 @@ sprintf_opcode_(CSOUND *csound,
         case 'g':
         case 'G':
 #ifdef HAVE_SNPRINTF
-          if(strlen(strseg) + 24 > maxChars) {
+          if(strlen(strseg) + 24 > (unsigned)maxChars) {
            int offs = outstring - str->data;
             str->data = mrealloc(csound, str->data,
                                  str->size  + 13);
@@ -378,7 +378,7 @@ sprintf_opcode_(CSOUND *csound,
             maxChars += 24;
             outstring = str->data + offs;
             //printf("maxchars = %d  %s\n", maxChars, strseg);
-	  }
+          }
           n = snprintf(outstring, maxChars, strseg, (double)*parm);
 #else
           n = sprintf(outstring, strseg, (double)*parm);
@@ -397,7 +397,7 @@ sprintf_opcode_(CSOUND *csound,
             maxChars += ((STRINGDAT*)parm)->size;
             outstring = str->data + offs;
           }
-	   n = sprintf(outstring, strseg, ((STRINGDAT*)parm)->data);
+          n = sprintf(outstring, strseg, ((STRINGDAT*)parm)->data);
           break;
         default:
           return StrOp_ErrMsg(p, "invalid format string");
@@ -409,7 +409,7 @@ sprintf_opcode_(CSOUND *csound,
             outstring = str->data + offs;
             str->size = maxChars*2;
             maxChars += str->size;
-           
+
         }
         outstring += n;
         len += n;
@@ -716,7 +716,7 @@ int strsub_opcode(CSOUND *csound, STRSUB_OP *p)
       end = tmp;
       rev = 1;
     }
-    
+
     src += strt;
     len = end - strt;
     if (UNLIKELY(len >=  p->Sdst->size)) {
